tests/wpt/mozilla/tests/bluetooth/getPrimaryServices/disconnect-called-during-with-uuid.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => {
        let promise = promise_rejects_dom(t, 'NetworkError', gattServer.getPrimaryServices(heart_rate.name));
        gattServer.disconnect();
        return promise;
    });
}, 'disconnect() called during getPrimaryServices. Reject with NetworkError.');
</script>
